requirements.txt
alembic==1.6.5
appnope==0.1.2
backcall==0.2.0
beautifulsoup4==4.9.3
certifi==2021.5.30
charset-normalizer==2.0.4
click==8.0.1
decorator==5.0.9
Flask==2.0.1
Flask-Migrate==3.1.0
Flask-SQLAlchemy==2.5.1
greenlet==1.1.1
idna==3.2
ipython==7.26.0
ipython-genutils==0.2.0
itsdangerous==2.0.1
jedi==0.18.0
Jinja2==3.0.1
Mako==1.1.4
MarkupSafe==2.0.1
matplotlib-inline==0.1.2
numpy==1.21.1
pandas==1.3.1
parso==0.8.2
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.19
psycopg2-binary==2.9.1
ptyprocess==0.7.0
Pygments==2.9.0
python-dateutil==2.8.2
python-dotenv==0.19.0
python-editor==1.0.4
pytz==2021.1
requests==2.26.0
six==1.16.0
soupsieve==2.2.1
SQLAlchemy==1.4.22
traitlets==5.0.5
urllib3==1.26.6
wcwidth==0.2.5
Werkzeug==2.0.1
